+++ CHANGELOG.md
@@ -9,11 +9,12 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 - EM pipeline starting from either halfmaps or a single map with optional blurring
 - EMDA map mask job
-- Servalcat trim, refine and NE map jobs
+- Servalcat trim, refine, FSC and NE map jobs
 - Refmac map to MTZ job
 - Libg job
 - Support for a user-supplied restraints dictionary for ligands
 - CCP-EM program tests
+- Renaming SUL residues to SO4 and HOH O1 atoms to O on reading a structure
 
 ### Changed